+++ models/Games.js
@@ -1,6 +1,5 @@
 const { Model, DataTypes } = require('sequelize');
 const sequelize = require('../config/connection.js');
-const { Games } = require('./index.js');
 
 class Games extends Model {}
