+++ views/login.handlebars
@@ -1,14 +1,3 @@
-{{!-- <!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Rankedify</title>
-</head> --}}
-
-{{!-- <body> --}}
     <div>
         <div>
             <h2>Login</h2>
@@ -49,6 +38,3 @@
     </div>
 
     <script src="./js/login.js"></script>
-{{!-- </body> --}}
-
-{{!-- </html> --}}
